% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.wrangle.R
\name{aes_assign}
\alias{aes_assign}
\title{Assign inputs to \code{x}, \code{y} or \code{dots}}
\usage{
aes_assign(data, x, y, ...)
}
\arguments{
\item{data, x, y, ...}{Arguments passed from \code{aes_loop()}}
}
\description{
\code{aes_assign()} figures out which columns/variables have been
passed and appropriatley assigns the columns/variables to their
respective mapping (\code{x}, \code{y}, or \code{dots}).
}
\details{
\code{aes_assign()} function is the first major function called
by \code{aes_loop()}.
}

