% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.wrangle.R
\name{aes_eval}
\alias{aes_eval}
\title{Assign inputs to \code{x}, \code{y} or \code{dots}}
\usage{
aes_eval(vars, x, y, dots)
}
\arguments{
\item{vars, x, y, ...}{Arguments passed from \code{aes_loop3()}.}
}
\description{
\code{aes_eval()} figures out which columns/variables have been
passed and appropriatley assigns the columns/variables to their
respective mapping (\code{x}, \code{y}, or \code{dots}).
}
\details{
\code{aes_eval()} function is the first major function called
by \code{aes_loop()}.
}

