% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.wrangle.R
\name{aes_group}
\alias{aes_group}
\title{Create unique pairings between \code{c(x, y)} and \code{dots}.}
\usage{
aes_group(lst)
}
\arguments{
\item{lst}{A list. The list that will be passed to \code{aes_group()}
will be the list produced by \code{aes_assing()}.}
}
\description{
\code{aes_group()} uses a list of \code{x}'s and \code{y}'s
to create each unique combination with \code{dots}.
}

