% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{expand.grid2}
\alias{expand.grid2}
\title{A new version of an old favorite with some extra options}
\usage{
expand.grid2(..., rm.dupes = TRUE, rm.dubs = TRUE)
}
\arguments{
\item{...}{Vectors to be expanded.}

\item{rm.dupes}{Removes duplicated "rows". If \code{TRUE} (default) then
rows that are unordered duplicates of other rows will be removed. i.e.
\code{c("A", "B", "C")} is the same as \code{c("C", "B", "A")}
 and any other combination of \code{"A"}, \code{"B"}, and \code{"C"}.}

\item{rm.dubs}{Removes a row in which all elements are the same. If
\code{TRUE} (default) then a row such as \code{c("A", "A", "A")} will be
removed.}
}
\description{
\code{expand.grid2()} creates a combination data frame from vectors or lists
but differs from the original \code{expand.grid()} in that it has two options
for removing two different type of duplicates. \code{stringsAsFactors} is
set to \code{TRUE}.
}

