% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.eval.R
\docType{data}
\name{fun.par}
\alias{fun.par}
\title{Regular expression pattern for determing if possible function parenthesis
are present. Searches for \code{"("} and \code{")"} preceeded by any number
of characters.}
\format{An object of class \code{character} of length 1.}
\usage{
fun.par
}
\description{
Regular expression pattern for determing if possible function parenthesis
are present. Searches for \code{"("} and \code{")"} preceeded by any number
of characters.
}
\keyword{datasets}

