% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggloop.R
\name{ggloop}
\alias{ggloop}
\title{Create ggplot plots in a loop.}
\usage{
ggloop(data, mappings = aes_loop(), remap_xy = TRUE, remap_dots = FALSE,
  gg_obs = TRUE, ..., environment = parent.frame())
}
\arguments{
\item{data}{Default dataset to use for plot. Must be a data frame and can be
only one data frame.}

\item{mappings}{List of aesthetic mappings to use for plots. Works like
\code{mapping} from \code{ggplot()}.}

\item{remap_xy}{Remapping behavior of the \code{x} and \code{y} vectors
specified in \code{aes_loop()}. See details below for more on remapping
behavior.}

\item{remap_dots}{Remapping behavior of the \code{...} vectors specified in
\code{aes_loop()}. See details below for more on remapping behavior.}

\item{gg_obs}{Logical. Specifies whether to return plots or the list (or
nested list) of aesthetics used to make such a plots.}

\item{...}{Other arguments. Similar to \code{ggplot()}'s \code{...}.}

\item{environment}{An environment and only one environment (cannot be a
vector). Similar to \code{ggplot()}'s \code{environment}.}
}
\description{
\code{ggloop()} mimics \code{ggplot()} by excepting a data frame and mappings
in order to return a plot. The main difference is that \code{ggloop()}
accepts vectors for aesthetics and returns a list or nested list of
\code{ggplot} plots.
}
\details{
\code{ggloop()} makes use of \code{aes_loop}, which is meant to
  mimic \code{aes} from \code{ggplot2}. Because of this, the remapping
  arguments are supplied to \code{ggloop} instead of \code{aes_loop()}.

The first remapping argument, \code{remap_xy} can take three values:
\itemize{ \item \code{TRUE} = The default behavior. All unqiue combinations
of \code{x} and \code{y} are generated. This means that if a variable (i.e.
\code{mpg}) is supplied in both \code{x} and \code{y}, then no mapping will
have \code{x} and \code{y} variables that are the same (i.e. \code{x -> mpg;
y -> mpg} will not ever happen). Likewise, no unordered pair duplicates will
happen (i.e. \code{x -> mpg; y -> cyl} and \code{x -> cyl; y -> mpg} will be
treated the same). \item \code{FALSE} = If \code{x} and \code{y} vectors are
not the same length, then the shorter of the two will be recycled. Recycling
is similar to \code{mapply()}'s recycling. \item \code{NA} = If \code{x} and
\code{y} vectors are not the same length, then the shorter of the two will
have \code{NA} assigned to the missing elements. These are meant to act as
placeholders during the wrangling operations (extracting and grouping the
aesthetics), and will be taken out before the final list of mappings is sent
to \code{ggloop()}. }
}

