% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%M\%}
\alias{\%M\%}
\title{The modified combination of the modulus function (\code{\%\%}) and
integer divisor function (\code{\%/\%}).}
\usage{
lhs \%M\% rhs
}
\arguments{
\item{lhs}{A number (integer or numeric)}

\item{rhs}{A number (integer or numeric)}
}
\description{
The placement of the arguments (\code{lhs} and \code{rhs}) does not matter
unlike the actual modulus function (\code{\%\%}) and integer divisor
function (\code{\%/\%})
}

