% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%||\%}
\alias{\%||\%}
\title{The replacement operator. Replaces the \code{lhs} with \code{rhs} on the
condition that \code{length(lhs) == FALSE} (the length is \code{0}).}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{An object of any length.}

\item{rhs}{A replacement value if  \code{length(lhs) == FALSE}.}
}
\description{
The replacement operator. Replaces the \code{lhs} with \code{rhs} on the
condition that \code{length(lhs) == FALSE} (the length is \code{0}).
}

