% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.eval.R
\name{is.fun}
\alias{is.fun}
\title{Is it a function?}
\usage{
is.fun(lst)
}
\arguments{
\item{lst}{A list of inputs wrapped in \code{substitute()} and coerced to a
list using \code{as.list()}.}
}
\description{
Attempts to decipher if a function other than \code{c()} has been supplied as
input. Returns the position of the possible non-\code{c} functions in
\code{lst}.
}

