% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.eval.R
\name{is.op}
\alias{is.op}
\title{Determine if an input uses an arithmetical operator (\code{/}, \code{+},
\code{-}, \code{*}, \code{^}).}
\usage{
is.op(lst)
}
\description{
Determine if an input uses an arithmetical operator (\code{/}, \code{+},
\code{-}, \code{*}, \code{^}).
}

