% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.eval.R
\name{is_dplyr}
\alias{is_dplyr}
\title{Determine if an input expression is a dplyr-like call.}
\usage{
is_dplyr(expr)
}
\arguments{
\item{expr}{A vector of substituted variables or functions. Either a call to
multiple variables with the outer \code{c()} wrapper stripped, a single
variable call wrapped in a \code{list()}, or a \code{:} expression.}
}
\description{
Determine if an input expression is a dplyr-like call.
}

