% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.eval.R
\name{is_op}
\alias{is_op}
\title{Determine if an input uses an arithmetical operator (\code{/}, \code{+},
\code{-}, \code{*}, \code{^}).}
\usage{
is_op(expr)
}
\arguments{
\item{expr}{An expression to be tested.}
}
\description{
Matches the arugment the \code{ops} string using \code{grep}. Any matches are
subsequently noted and the unique list is returned.
}

