% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{isFALSE}
\alias{isFALSE}
\title{This is an abbreviation of identical(FALSE, x) to go along with isTRUE()}
\usage{
isFALSE(x)
}
\description{
Use this when needing to test explicitly if a value is FALSE.
}

