% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.eval.R
\docType{data}
\name{ops}
\alias{ops}
\title{Arithmetic operators to search for.}
\format{An object of class \code{character} of length 5.}
\usage{
ops
}
\description{
Arithmetic operators to search for.
}
\keyword{datasets}

