% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.remap.R
\name{remap_xy_FALSE}
\alias{remap_xy_FALSE}
\title{Mimicks R's internal recycling mechanism for the shorter of the two vectors.}
\usage{
remap_xy_FALSE(lst)
}
\arguments{
\item{lst}{A list. The list passe will be the raw list generated from calling
\code{aes_assign()} and is ran before a remap function for the any "dots"
in the list.}
}
\description{
The smallest of the two vectors (\code{x} or \code{y}) will be recycled in
a manner similar to R's internal recycling mechanism.
}

