% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.remap.R
\name{remap_xy_NA}
\alias{remap_xy_NA}
\title{Attaches NA during recycling of the smaller of the two vectors.}
\usage{
remap_xy_NA(lst)
}
\arguments{
\item{lst}{A list. The list passe will be the raw list generated from calling
\code{aes_assign()} and is ran before a remap function for the any "dots"
in the list.}
}
\description{
The smallest of the two vectors (\code{x} or \code{y}) will be recycled with
NA instead of using the vector itself (similar to R's internal recycling
mechanism).
}

