% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.eval.R
\name{rm.gg2}
\alias{rm.gg2}
\title{Remove \code{ggplot2} style and stand-alone aesthetic arguments (i.e.
\code{y}, \code{x:z}, etc).}
\usage{
rm.gg2(x)
}
\arguments{
\item{x}{A parse tree generated by \code{substitute()}. If it is not wrapped
by \code{c()} (just a single expression) then it is advised to wrap \code{x}
with \code{list()}.}
}
\description{
Remove \code{ggplot2} style and stand-alone aesthetic arguments (i.e.
\code{y}, \code{x:z}, etc).
}

