% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.stuff.R
\docType{data}
\name{select_helpers}
\alias{select_helpers}
\title{Helper functions to select variable names using non-standard evaluation.}
\format{An object of class \code{list} of length 7.}
\usage{
select_helpers
}
\description{
Helper functions to select variable names using non-standard evaluation.
}
\seealso{
Source for \code{select_helpers} and the helper functions can be found at
\href{https://github.com/hadley/dplyr/blob/master/R/select-utils.R}{~/dplyr/R/select-vars.R}
and
\href{https://github.com/hadley/dplyr/blob/master/R/select-utils.R}{~/dplyr/R/select-utils.R}.
}
\keyword{datasets}

