% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{what}
\alias{what}
\title{Console function for determing: class, type, mode, and names of an object.}
\usage{
what(x, SIMPLIFY = TRUE)
}
\arguments{
\item{x}{An object.}

\item{SIMPLIFY}{Option to simplify result to a
vector (deafult is \code{TRUE}). Result is a list if \code{FALSE}.}
}
\description{
Console function for determing: class, type, mode, and names of an object.
}

