% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.loop.R
\name{aes_loop}
\alias{aes_loop}
\title{Create a grouped list of aesthetic mappings to pass to \code{ggloop()}.}
\usage{
aes_loop(x, y, ...)
}
\arguments{
\item{x, y, ...}{A vector of variable names. Vector can consist of a combination of
\code{dplyr}-like symbols (unqouted names) and \code{numerics/integers}
referencing the variable position within the data frame assigned to
\code{data}. \code{...} arguments (of course) must have an argument name.}
}
\description{
This version of \code{aes_loop} is to be used inside \code{ggloop()}. If
you wish to return the grouped list of uneval aesthetics, then use
\code{aes_loop2()}.
}
\details{
\code{aes_loop()} is meant solely to be called within \code{ggloop()}. To
create the raw list of grouped mappings, use \code{aes_loop2()}.
}

