% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.eval.R
\name{is.c}
\alias{is.c}
\title{Determine if the first element of a parse tree is identical to the \code{c}
function.}
\usage{
is.c(x)
}
\arguments{
\item{x}{A parse tree generated by \code{substitute()}.}
}

