% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.map.R
\name{map_aes}
\alias{map_aes}
\title{Loop through a list an assing "uneval" class to each parsed element.}
\usage{
map_aes(lst)
}
\description{
This is essentially the core of \code{aes()} from \code{ggplot2} placed
inside of an \code{lapply()} loop. This function will be called using an
\code{mapply()} loop.
}

