% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiplot}
\alias{multiplot}
\title{Plot multiple plots in one layout.}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Unrendered \code{ggplot} plots tp view in a single pane.}

\item{plotlist}{A list of \code{ggplot} plots (if you prefer).}

\item{file}{Not yet known/defined.}

\item{cols}{Number of columns in the layout.}

\item{layout}{A matrix specifying the layout. If present, \code{cols} is
ignored.}
}
\description{
\code{ggplot} objects can be passed in ..., or to plotlist (as a list of
ggplot objects) - cols:   Number of columns in layout - layout: A matrix
specifying the layout. If present, 'cols' is ignored.
}
\details{
If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE), then
plot 1 will go in the upper left, 2 will go in the upper right, and 3 will go
all the way across the bottom.
}
\seealso{
The source for the code is Winston Chang's "Cookbook for R", and can be found
at the following link:
\url{http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}

A list of StackOverflow Q&A's:
\itemize{
\item \href{http://stackoverflow.com/questions/13788563/r-multiplot-not-working-because-of-error-in-usemethodgetmodelinfo-model}{R: multiplot not working because of error in UseMethod(“getModelInfo”,model)}
\item \href{http://stackoverflow.com/questions/11721401/r-save-multiplot-to-file}{R - save multiplot to file}
\item \href{http://stackoverflow.com/questions/8615530/place-title-of-multiplot-panel-with-ggplot2}{Place title of multiplot panel with ggplot2}
\item \href{http://stackoverflow.com/questions/28138257/r-multiplot-write-plot-to-file-grid-package}{R multiplot write plot to file grid package}
}
}

