% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.map.R
\name{name_groups}
\alias{name_groups}
\title{Extract names for the first level of list components for the returned value
of \code{ggloop()}.}
\usage{
name_groups(lst, dots.vector)
}
\description{
Extract names for the first level of list components for the returned value
of \code{ggloop()}.
}

